requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
constantly==15.1.0
coverage==6.3.1
cryptography==36.0.1
cssselect2==0.5.0
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.4
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-oauth-toolkit==1.7.0
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
filelock==3.6.0
future==0.18.2
fuzzywuzzy==0.18.0
hashring==1.5.1
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.1
incremental==21.3.0
jwcrypto==1.0
lxml==4.8.0
Markdown==3.3.6
memcache==0.11.0
msgpack==1.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.1
oauthlib==3.2.0
oscrypto==1.3.0
phonenumbers==8.12.46
phonenumberslite==8.12.46
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
virtualenv==20.13.2
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.7
zipp==3.7.0
zope.interface==5.4.0
